Add USGS Details page
The value in each page is served via an iframe. This simple HTML
page will fetch values corresponding to a "code" query parameter
and populate them in elements. In case of any errors, the page
remains blank and user will not see anything. Warnings will be
logged to the console.

Styling for this page has been taken from the sample HTML pages
serving as placeholders previously. This is one of them:
http://www.pixelpresto.com/Dispatches/test-content-Temperature.html