Use listener factory for given metric
Instead of having one fixed event listener, we now configure it by
passing the appropriate metric as a parameter. Thus, the new
listener function takes a metric and returns the true listener
function. It is an application of the Factory pattern.

We calculate the fallback value for the current month, and the
units, for the given metric. In case the API request fails, we use
these fallback values instead.